Add flag to force `git reset --hard`
e.g. to update ~ crontab.
This will discard uncommitted changes though; maybe add a check that the diff is -1/+1?

+++ install.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+
+# Handle -y flag to 
+force_git_reset=0
+
+while getopts 'y' opt; do
+    case $opt in 
+        y) force_git_reset=1 ;;
+    esac
+done
+
 makepkg -f
 ls -t nemo-ybnd-personal* | head -1 | xargs -I $ sudo pacman -U --noconfirm $
 sudo rm -r nemo
@@ -14,12 +24,16 @@ echo $'\n'
 
 while [ 1 ]
 do
-	read -p $'git reset --hard? (y/n) ' -n 1 choice
+    if [ $force_git_reset -eq 1 ]; then
+        choice="y"
+    else
+        read -p $'git reset --hard? (y/n) ' -n 1 choice
+    fi	
 
 	case "$choice" in 
 		y|Y )
 			echo $'\n'
-			git reset --hard
+			git reset --hard			
 			break
 			;;
 		n|N )
